* migrated the project to Maven; * small refactoring; * updated libraries; * changed the project Java from 7th to 8th; * divided sources to "main" & "test".
