--- OAuth2Client_Java/src/com/sap/security/oa2/test/saml.properties
@@ -1,20 +1,0 @@
-# Keystore Path
-ks_resource=/com/sap/security/oa2/test/sts.jks
-# Keystore Type
-ks_type=JKS
-# Keystore Password. Same password is used to open the keystore and key entry
-ks_pwd=abcd1234
-# Keystore key alias
-ks_alias=STS
-# SAML target audience (->name of local provider from transaction SAML2)
-saml_audience_restriction=SIM_000
-# SAML issuer name
-saml_issuer=DemoAppIDP
-# SAML name
-saml_session_authentication=urn:none
-# OAuth2 client id
-oa2_client_id=LEAVEAPP
-# OAuth2 client password
-oa2_client_secret=abcd1234
-# OAuth2 Token endpoint
-oa2_token_endpoint=https://dewdfgwd00214.wdf.sap.corp:44318/sap/bc/sec/oauth2/token
