+++ OAuth2Client_Java/src/main/java/com/sap/security/oa2/AccessTokenException.java
@@ -1,0 +1,13 @@
+package com.sap.security.oa2;
+
+public class AccessTokenException extends Exception {
+
+    public AccessTokenException(Exception ex) {
+	super(ex);
+    }
+
+    public AccessTokenException(String message) {
+	super(message);
+    }
+
+}
