+++ OAuth2Client_Java/src/main/java/com/sap/security/oa2/MissingPropertyException.java
@@ -1,0 +1,7 @@
+package com.sap.security.oa2;
+
+public class MissingPropertyException extends Exception {
+    public MissingPropertyException(String missingProperty) {
+	super("Missing property: " + missingProperty);
+    }
+}
