+++ OAuth2Client_Java/src/main/java/com/sap/security/oa2/SamlTokenFactory.java
@@ -1,0 +1,10 @@
+package com.sap.security.oa2;
+
+import java.util.Properties;
+
+
+public interface SamlTokenFactory {
+
+    public abstract String getSamlAssertion(Properties cfgProperties) throws SAMLException;
+
+}
